+++ src/Model/AddressOsGridRefInterface.php
@@ -6,13 +6,11 @@ namespace LocalgovDrupal\OsPlacesGeocoder\Model;
 
 /**
  * Ordnance Survey Location record interface.
- *
- * Adds Ordnance Survey Grid reference support to UPRN-based Location records.
  */
-interface LocationOsPlacesInterface extends LocationUprnInterface {
+interface AddressOsGridRefInterface {
 
   /**
-   * Getter for the OS grid reference object.
+   * Returns the OS grid reference object.
    */
   public function getOsGridRef() :OsGridRef;
