+++ src/Model/OsPlacesAddress.php
@@ -4,13 +4,53 @@ declare(strict_types = 1);
 
 namespace LocalgovDrupal\OsPlacesGeocoder\Model;
 
+use Geocoder\Model\Address;
+
 /**
  * Ordnance Survey places address record.
  *
  * Includes Ordnance Survey Grid reference in addition to everything a
  * UPRN-based address record would have.
  */
-final class OsPlacesAddress extends UprnAddress implements LocationOsPlacesInterface {
+final class OsPlacesAddress extends Address implements AddressOsGridRefInterface, AddressUprnInterface {
+
+  /**
+   * Unique Property Reference Number.
+   *
+   * @var string
+   *
+   * @see https://en.wikipedia.org/wiki/Unique_Property_Reference_Number
+   */
+  protected $uprn = '';
+
+  /**
+   * The full address in one line.
+   *
+   * @var string
+   */
+  protected $displayName = '';
+
+  /**
+   * Flat number.
+   *
+   * @var string
+   */
+  protected $flat = '';
+
+  /**
+   * House name.
+   *
+   * @var string
+   */
+  protected $houseName = '';
+
+  /**
+   * Organisation.
+   *
+   * @var string
+   */
+  protected $org = '';
+
 
   /**
    * All numeric Ordnance Survey National Grid reference.
@@ -30,6 +70,46 @@ final class OsPlacesAddress extends UprnAddress implements LocationOsPlacesInter
   }
 
   /**
+   * {@inheritdoc}
+   */
+  public function getUprn() :string {
+
+    return $this->uprn;
+  }
+
+  /**
+   * The full address, in one line.
+   */
+  public function getDisplayName() :string {
+
+    return $this->displayName;
+  }
+
+  /**
+   * Returns the flat number if any.
+   */
+  public function getFlat() :string {
+
+    return $this->flat;
+  }
+
+  /**
+   * Returns the house name if any.
+   */
+  public function getHouseName() :string {
+
+    return $this->houseName;
+  }
+
+  /**
+   * Returns the organisation name if any.
+   */
+  public function getOrganisationName() :string {
+
+    return $this->org;
+  }
+
+  /**
    * Creates an Address from an array.
    *
    * @return static
@@ -42,11 +122,17 @@ final class OsPlacesAddress extends UprnAddress implements LocationOsPlacesInter
       $self->osGridRef = new OsGridRef((int) $data['easting'], (int) $data['northing']);
     }
 
+    $self->uprn        = $data['uprn'] ?? '';
+    $self->displayName = $data['display'] ?? '';
+    $self->flat        = $data['flat'] ?? '';
+    $self->houseName   = $data['houseName'] ?? '';
+    $self->org         = $data['org'] ?? '';
+
     return $self;
   }
 
   /**
-   * Appends uprn and display value to Location array.
+   * Appends osgrid, uprn and display value to Location array.
    *
    * {@inheritdoc}
    */
@@ -58,6 +144,12 @@ final class OsPlacesAddress extends UprnAddress implements LocationOsPlacesInter
       $array['easting']  = $this->osGridRef->getEasting();
       $array['northing'] = $this->osGridRef->getNorthing();
     }
+
+    $array['uprn']      = $this->getUprn();
+    $array['display']   = $this->getDisplayName();
+    $array['flat']      = $this->getFlat();
+    $array['houseName'] = $this->getHouseName();
+    $array['org']       = $this->getOrganisationName();
 
     return $array;
   }
