--- src/Model/UprnAddress.php
@@ -1,129 +1,0 @@
-<?php
-
-declare(strict_types = 1);
-
-namespace LocalgovDrupal\OsPlacesGeocoder\Model;
-
-use Geocoder\Model\Address;
-
-/**
- * UPRN-based address record.
- *
- * UPRN = Unique Property Reference Number.
- */
-class UprnAddress extends Address implements LocationUprnInterface {
-
-  /**
-   * Unique Property Reference Number.
-   *
-   * @var string
-   *
-   * @see https://en.wikipedia.org/wiki/Unique_Property_Reference_Number
-   */
-  protected $uprn = '';
-
-  /**
-   * The full address in one line.
-   *
-   * @var string
-   */
-  protected $displayName = '';
-
-  /**
-   * Flat number.
-   *
-   * @var string
-   */
-  protected $flat = '';
-
-  /**
-   * House name.
-   *
-   * @var string
-   */
-  protected $houseName = '';
-
-  /**
-   * Organisation.
-   *
-   * @var string
-   */
-  protected $org = '';
-
-  /**
-   * As it says on the tin.
-   */
-  public function getUprn() :string {
-
-    return $this->uprn;
-  }
-
-  /**
-   * The full address, in one line.
-   */
-  public function getDisplayName() :string {
-
-    return $this->displayName;
-  }
-
-  /**
-   * Returns the flat number if any.
-   */
-  public function getFlat() :string {
-
-    return $this->flat;
-  }
-
-  /**
-   * Returns the house name if any.
-   */
-  public function getHouseName() :string {
-
-    return $this->houseName;
-  }
-
-  /**
-   * Returns the organisation name if any.
-   */
-  public function getOrganisationName() :string {
-
-    return $this->org;
-  }
-
-  /**
-   * Creates an Address from an array.
-   *
-   * @return static
-   */
-  public static function createFromArray(array $data) {
-
-    $self = parent::createFromArray($data);
-
-    $self->uprn        = $data['uprn'] ?? '';
-    $self->displayName = $data['display'] ?? '';
-    $self->flat        = $data['flat'] ?? '';
-    $self->houseName   = $data['houseName'] ?? '';
-    $self->org         = $data['org'] ?? '';
-
-    return $self;
-  }
-
-  /**
-   * Appends uprn and display value to Location array.
-   *
-   * {@inheritdoc}
-   */
-  public function toArray() :array {
-
-    $array = parent::toArray();
-
-    $array['uprn']      = $this->getUprn();
-    $array['display']   = $this->getDisplayName();
-    $array['flat']      = $this->getFlat();
-    $array['houseName'] = $this->getHouseName();
-    $array['org']       = $this->getOrganisationName();
-
-    return $array;
-  }
-
-}
